25 warnings remaining
Former-commit-id: d5bfdd1523243c6727bf8ed1c1b824c5c39421fb

+++ tools/pg_psql/copy.cpp
@@ -450,7 +450,7 @@ handleCopyOut(PGconn *conn, FILE *copystream, PGresult **res)
 
 		if (buf)
 		{
-			if (OK && fwrite(buf, 1, ret, copystream) != ret)
+			if (OK && fwrite(buf, 1, ret, copystream) != static_cast<unsigned int>(ret))
 			{
 				psql_error("could not write COPY data: %s\n",
 						   strerror(errno));
